@@ -18,12 +18,13 @@
     }
   ],
   "require":{
-    "php":">=7.1",
-    "contao/core-bundle":"^4.4",
+    "php":"^7.4 || ^8.0",
+    "contao/core-bundle":"^4.9",
     "contao-estatemanager/core": "^1.0"
   },
   "require-dev": {
-    "contao/manager-plugin": "^2.0"
+    "contao/manager-plugin": "^2.0",
+    "contao/easy-coding-standard": "^3.4"
   },
   "conflict": {
     "contao/core": "*",
@@ -45,5 +46,10 @@
   },
   "extra":{
     "contao-manager-plugin": "ContaoEstateManager\\Video\\ContaoManager\\Plugin"
+  },
+  "scripts": {
+    "cs-fixer": [
+      "vendor/bin/ecs check src/ --fix --ansi"
+    ]
   }
 }
