@@ -1,11 +1,14 @@
 <?php
-/**
+
+declare(strict_types=1);
+
+/*
  * This file is part of Contao EstateManager.
  *
- * @link      https://www.contao-estatemanager.com/
- * @source    https://github.com/contao-estatemanager/video
- * @copyright Copyright (c) 2019  Oveleon GbR (https://www.oveleon.de)
- * @license   https://www.contao-estatemanager.com/lizenzbedingungen.html
+ * @see        https://www.contao-estatemanager.com/
+ * @source     https://github.com/contao-estatemanager/video
+ * @copyright  Copyright (c) 2021 Oveleon GbR (https://www.oveleon.de)
+ * @license    https://www.contao-estatemanager.com/lizenzbedingungen.html
  */
 
 namespace ContaoEstateManager\Video;
@@ -13,7 +16,6 @@ namespace ContaoEstateManager\Video;
 use Contao\BackendTemplate;
 use ContaoEstateManager\ExposeModule;
 use ContaoEstateManager\Translator;
-use Patchwork\Utf8;
 
 /**
  * Expose module "video".
@@ -23,42 +25,44 @@ use Patchwork\Utf8;
 class ExposeModuleVideo extends ExposeModule
 {
     /**
-     * Template
+     * Template.
+     *
      * @var string
      */
     protected $strTemplate = 'expose_mod_video';
 
     /**
-     * Do not display the module if there are no real estates
+     * Do not display the module if there are no real estates.
      *
      * @return string
      */
     public function generate()
     {
-        if (TL_MODE == 'BE')
+        if (TL_MODE === 'BE')
         {
             $objTemplate = new BackendTemplate('be_wildcard');
-            $objTemplate->wildcard = '### ' . Utf8::strtoupper($GLOBALS['TL_LANG']['FMD']['video'][0]) . ' ###';
+            $objTemplate->wildcard = '### '.mb_strtoupper($GLOBALS['TL_LANG']['FMD']['video'][0], 'UTF-8').' ###';
             $objTemplate->title = $this->headline;
             $objTemplate->id = $this->id;
             $objTemplate->link = $this->name;
-            $objTemplate->href = 'contao/main.php?do=expose_module&amp;act=edit&amp;id=' . $this->id;
+            $objTemplate->href = 'contao/main.php?do=expose_module&amp;act=edit&amp;id='.$this->id;
 
             return $objTemplate->parse();
         }
 
         $strBuffer = parent::generate();
-        return $this->isEmpty && !!$this->hideOnEmpty ? '' : $strBuffer;
+
+        return $this->isEmpty && (bool) $this->hideOnEmpty ? '' : $strBuffer;
     }
 
     /**
-     * Generate the module
+     * Generate the module.
      */
-    protected function compile()
+    protected function compile(): void
     {
         $arrLinks = Video::collectVideoLinks($this->realEstate->links, 1);
 
-        if($arrLinks === null)
+        if (null === $arrLinks)
         {
             $this->isEmpty = true;
         }
@@ -68,18 +72,18 @@ class ExposeModuleVideo extends ExposeModule
             $link = $arrLinks[0];
 
             // generate link with attributes
-            $settings = array(
-                'autoplay'   => 1, // ToDo: Determine, if field should be added to expose module
-                'controls'   => 1, // ToDo: Add field to expose module
+            $settings = [
+                'autoplay' => 1, // ToDo: Determine, if field should be added to expose module
+                'controls' => 1, // ToDo: Add field to expose module
                 'fullscreen' => 1, // ToDo: Add field to expose module
-            );
+            ];
 
             $link = Video::generateAttributeLink($link, $settings);
 
             // set template information
             $this->Template->link = $link;
         }
-        
+
         $this->Template->label = Translator::translateExpose('button_video');
     }
 }
